# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CommsPro - Professional Communication Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: getUsers
      tags: [users]
      summary: List all users
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ─── CONTACTS ───────────────────────────────────────────────────────────────
...
  /whatsapp/conversations/{phone}/messages:
    get:
      operationId: getWhatsappMessages
      tags: [whatsapp]
      summary: Get messages in a conversation
      parameters:
        - name: phone
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WhatsappMessage"
  /whatsapp/send:
    post:
      operationId: sendWhatsapp
      tags: [whatsapp]
      summary: Send a WhatsApp message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WhatsappInput"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WhatsappMessage"

  # ─── AI AGENTS ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string

    # ── Auth ──────────────────────────────────────────────────────────────────
...
    User:
      type: object
      required: [id, name, email, role, isOnline, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, agent, viewer]
        isOnline:
          type: boolean
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    UserInput:
      type: object
      required: [name, email, role, password]
      properties:
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, agent, viewer]
        password:
          type: string
    UserUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, agent, viewer]
        isOnline:
          type: boolean

    # ── Contact ───────────────────────────────────────────────────────────────
...
    WhatsappMessage:
      type: object
      required: [id, direction, from, to, body, status, createdAt]
      properties:
        id:
          type: integer
        twilioSid:
          type: ["string", "null"]
        direction:
          type: string
          enum: [inbound, outbound]
        from:
          type: string
        to:
          type: string
        body:
          type: string
        status:
          type: string
          enum: [queued, sending, sent, delivered, failed, read]
        mediaUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    WhatsappInput:
      type: object
      required: [to, body]
      properties:
        to:
          type: string
        body:
          type: string
        mediaUrl:
          type: string

    # ── AI Agent ──────────────────────────────────────────────────────────────
...